requirements.txt
selenium
beautifulsoup4
schedule
discord-webhook
psycopg2-binary
python-dotenv==1.0.1
